fix deadlock in ppage_dcache_open
We can't dput dentry when iterating parent's subdir
list because dput will get parent's dentry lock.
To resolve this issue, we add a struct list_head in
p_info so that we can add the inode of dentry to
delete to a list. We delete those dentries after
finishing iterating the parent's subdir list.